temporary save. Camera fixes
