+++ WebCam.py
@@ -1,10 +1,18 @@
-def getWebCamImage(camera):
+camera = None
+
+def getImage():
 	retval, image = camera.read()
 	return image
 
-def initializeWebCamera(resolution):
+def initializeCamera(resolution=(640,360)):
+	import cv2
 	camera = cv2.VideoCapture(0)
-	camera.set(3, resolution[0])
-	camera.set(4, resolution[1])
-	# camera.set(15, 0.1) # exposure
+	setCamera(resolution=resolution)
+
+def setCamera(resolution=False, exposure=False)
+	if resolution:
+		camera.set(3, resolution[0])
+		camera.set(4, resolution[1])
+	if exposure
+		camera.set(15, 0.1) # exposure
 	return camera
